feat: Journal Instances
Journal view pane now displays title instances along with associated identifiers when available
