--- src/components/JournalSections/JournalInfo/index.js
@@ -1,1 +1,0 @@
-export { default } from './JournalInfo';
